+++ doc/changelog.rst
@@ -74,6 +74,7 @@ Localization
 Admin
 ~~~~~
 
+- Add admin module for managing Django permission groups
 - Add `get_required_permission` to ``AdminModule``
 - Add support for module-level permissions
 - Remove Select2Multiple-widget as unused